fix(TokenAllowance): Crash
